backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.0
python-dotenv==1.0.0
Werkzeug==2.3.0
